requirements.txt
Flask
Werkzeug
python-dotenv
FLask-SQLAlchemy
flask_cors